sxmo config update

+++ sxmo/hooks/sxmo_hook_ring.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# configversion: c582971c48f67bb038dd16bbbec081fa
+# configversion: fc857f0eec65ce76ee07251621c457f4
 # SPDX-License-Identifier: AGPL-3.0-only
 # Copyright 2022 Sxmo Contributors
 
@@ -7,6 +7,9 @@
 # You can use it to play a ring tone
 
 # $1 = Contact Name or Number (if not in contacts)
+
+# shellcheck source=scripts/core/sxmo_common.sh
+. sxmo_common.sh
 
 # Only vibrate if you already got an active call
 if sxmo_modemcall.sh list_active_calls \
@@ -23,15 +26,44 @@ if ! sxmo_modemcall.sh list_active_calls \
 	exit
 fi
 
-# Start the mpv ring until another hook kill it or the max (10) is reached
-mpv --quiet --no-video --loop=10 /usr/share/sxmo/ring.ogg &
-MPVID=$!
-echo "$MPVID" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
-
 ~/dotfiles/notifysend.sh notify/proycon/phone "$1" &
 
-# Vibrate while mpv is running
-while kill -0 $MPVID; do
+# RING & VIBRATE MODE (DEFAULT)
+if [ ! -f "$XDG_CONFIG_HOME"/sxmo/.noring ] && [ ! -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ]; then
+	sxmo_log "RING AND VIBRATE"
+
+	# In order for this to work, you will need to install playerctl and run playerctld
+	# In order for this to work with mpv, you will need to install mpv-mdis.
+	sxmo_playerctl.sh pause_all
+
+	timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
+		--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
+	MPVID=$!
+	echo "$MPVID" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+	# vibrate while mpv is running
+	while kill -0 $MPVID; do
 		sxmo_vibrate 1500
 		sleep 0.5
-done
+	done
+
+# RING-ONLY MODE
+elif [ ! -f "$XDG_CONFIG_HOME"/sxmo/.noring ] && [ -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ]; then
+	sxmo_log "RING ONLY"
+
+	# In order for this to work, you will need to install playerctl and run playerctld
+	# In order for this to work with mpv, you will need to install mpv-mdis.
+	sxmo_playerctl.sh pause_all
+
+	timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
+		--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
+	echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+
+# VIBRATE-ONLY MODE
+elif [ ! -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ] && [ -f "$XDG_CONFIG_HOME"/sxmo/.noring ]; then
+	smxo_log "VIBRATE ONLY"
+	for _ in $(seq 5); do
+		sxmo_vibrate 1500
+		sleep 0.5
+	done &
+	echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+fi
